backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Google APIs
google-analytics-data==0.17.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0

# Azure SDK (for later)
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Utilities
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0